refactor: streamline navigation and integrate Navbar component across multiple detection pages
