+++ src/locales/ta.json
@@ -1,7 +1,8 @@
 {
   "common": {
     "company_name": "iPaddyCare",
-    "company_description": "ஸ்மார்ட் விவசாய தொழில்நுட்பத்துடன் இலங்கை விவசாயிகளை வலுப்படுத்துதல்."
+    "company_description": "ஸ்மார்ட் விவசாய தொழில்நுட்பத்துடன் இலங்கை விவசாயிகளை வலுப்படுத்துதல்.",
+    "back_to_home": "முகப்புக்கு திரும்பு"
   },
   "navigation": {
     "features": "அம்சங்கள்",
